requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary>=2.9.10
python-dotenv==1.0.0
pandas>=2.2.0
openpyxl==3.1.2
xhtml2pdf>=0.2.16
Jinja2==3.1.2

